+++ database/DbRule.cpp
@@ -20,7 +20,7 @@ static int RuleParse(sqlite3_stmt *stmt, void *ptr)
 				string data = Util::setString(reinterpret_cast<const char *>(sqlite3_column_text(stmt, index++)));
 				int type = sqlite3_column_int(stmt, index++);
 				bool enable = sqlite3_column_blob(stmt, index++);
-				int addr = sqlite3_column_int(stmt, index++);
+				uint16_t addr = sqlite3_column_int(stmt, index++);
 				string name = Util::setString(reinterpret_cast<const char *>(sqlite3_column_text(stmt, index++)));
 				string ruledata;
 				string decode = macaron::Base64::Decode(data, ruledata);
