+++ device/ble/DeviceBleDoorSensor.cpp
@@ -1,7 +1,7 @@
 #include "DeviceBleDoorSensor.h"
 #include "Log.h"
 
-DeviceBleDoorSensor::DeviceBleDoorSensor(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint16_t version)
+DeviceBleDoorSensor::DeviceBleDoorSensor(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint16_t version)
 		: DeviceBle(id, name, mac, dataJson, addr, BLE_DOOR_SENSOR, version)
 {
 	moduleDoorHangOn = new ModuleDoorHangOn(this, addr);
