+++ device/ble/DeviceBleLightOnoffCctDim.cpp
@@ -1,7 +1,7 @@
 #include "DeviceBleLightOnoffCctDim.h"
 #include "Log.h"
 
-DeviceBleLightOnoffCctDim::DeviceBleLightOnoffCctDim(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version)
+DeviceBleLightOnoffCctDim::DeviceBleLightOnoffCctDim(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version)
 	: DeviceBle(id, name, mac, dataJson, addr, type, version)
 {
 	moduleOnOff = new ModuleOnOff(this, addr, KEY_ATTRIBUTE_ONOFF);
