+++ device/ble/DeviceBlePirLightSensorDC.h
@@ -19,5 +19,5 @@ private:
 	ModuleTimeActionPir *moduleTimeActionPir;
 
 public:
-	DeviceBlePirLightSensorDC(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint16_t version);
+	DeviceBlePirLightSensorDC(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint16_t version);
 };
