+++ device/ble/DeviceBleScreenTouch.cpp
@@ -5,7 +5,7 @@
 #include "Log.h"
 #include "Http.h"
 
-DeviceBleScreenTouch::DeviceBleScreenTouch(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint16_t version)
+DeviceBleScreenTouch::DeviceBleScreenTouch(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint16_t version)
 	: DeviceBle(id, name, mac, dataJson, addr, BLE_AC_SCENE_SCREEN_TOUCH, version)
 {
 	moduleNotifyScene = new ModuleNotifyScene(this, addr);
