+++ device/ble/DeviceBleSensorPm.h
@@ -14,5 +14,5 @@ private:
 	ModuleTempHum *moduleTempHum;
 
 public:
-	DeviceBleSensorPm(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint16_t version);
+	DeviceBleSensorPm(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint16_t version);
 };
