+++ device/ble/DeviceBleSensorTempHum.cpp
@@ -1,7 +1,7 @@
 #include "DeviceBleSensorTempHum.h"
 #include "Log.h"
 
-DeviceBleSensorTempHum::DeviceBleSensorTempHum(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint16_t version)
+DeviceBleSensorTempHum::DeviceBleSensorTempHum(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint16_t version)
 		: DeviceBle(id, name, mac, dataJson, addr, BLE_TEMP_HUM_SENSOR, version)
 {
 	moduleTempHum = new ModuleTempHum(this, addr);
