+++ device/ble/DeviceBleSwitchElectrical.cpp
@@ -1,7 +1,7 @@
 #include "DeviceBleSwitchElectrical.h"
 #include "Log.h"
 
-DeviceBleSwitchElectrical::DeviceBleSwitchElectrical(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version, uint8_t countElement)
+DeviceBleSwitchElectrical::DeviceBleSwitchElectrical(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version, uint8_t countElement)
 		: DeviceBle(id, name, mac, dataJson, addr, type, version)
 {
 	this->countElement = countElement;
