+++ device/ble/DeviceBleSwitchScene6DC.h
@@ -13,5 +13,5 @@ private:
 	ModuleBatteryLevel *moduleBatteryLevel;
 
 public:
-	DeviceBleSwitchScene6DC(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version);
+	DeviceBleSwitchScene6DC(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version);
 };
