+++ device/ble/DeviceBleSwitchTouch.h
@@ -8,5 +8,5 @@ class DeviceBleSwitchTouch : public DeviceBle
 {
 private:
 public:
-	DeviceBleSwitchTouch(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version, uint8_t numRelay = 1);
+	DeviceBleSwitchTouch(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version, uint8_t numRelay = 1);
 };
