+++ device/ble/DeviceBleSwitchTouchRgb.h
@@ -8,5 +8,5 @@ class DeviceBleSwitchTouchRgb : public DeviceBle
 {
 private:
 public:
-	DeviceBleSwitchTouchRgb(string id, string name, string mac, Json::Value &dataJson, uint32_t addr, uint32_t type, uint16_t version, uint8_t countElement);
+	DeviceBleSwitchTouchRgb(string id, string name, string mac, Json::Value &dataJson, uint16_t addr, uint32_t type, uint16_t version, uint8_t countElement);
 };
