+++ device/ble/module/ModuleDim.h
@@ -9,7 +9,7 @@ protected:
 	uint16_t dim;
 
 public:
-	ModuleDim(Device *device, uint32_t addr);
+	ModuleDim(Device *device, uint16_t addr);
 	~ModuleDim();
 
 #ifdef CONFIG_SAVE_ATTRIBUTE
